Moving the reread file check to base scheduler and some cleanup
